displayed existing blogs on homepage, edited homeroutes, blog and comment model dates
