+++ controllers/api/index.js
@@ -5,7 +5,7 @@ const blogpostRoutes = require('./blogpostRoutes');
 const commentsRoutes = require('./commentsRoutes');
 
 router.use('/users', usersRoutes);
-// router.use('/blogposts', blogpostRoutes);
+router.use('/blogposts', blogpostRoutes);
 // router.use('/comments', commentsRoutes);
 
 module.exports = router;
