+++ models/Comments.js
@@ -16,7 +16,7 @@ Comments.init(
             allowNull: false
         },
         date: {
-            type: DataTypes.DATE,
+            type: DataTypes.DATEONLY,
             allowNull: false,
             defaultValue: DataTypes.NOW
         },
