+++ utils/auth.js
@@ -1,0 +1,9 @@
+const withAuth = (req, res, next) => {
+    if (!req.session.logged_in) {
+      res.redirect('/login');
+    } else {
+      next();
+    }
+  };
+  
+  module.exports = withAuth;
