+++ views/homepage.handlebars
@@ -4,8 +4,27 @@
     </p>
 </section>
 
-<section class="blog-list is medium">
-    {{#each allBlogs as |blog|}}
-        
-    {{/each}}
-</section>
+<div class="columns">
+    <div class="column is-8 is-offset-1">
+        <section class="blog-list">
+            {{#each allBlogs as |blog|}}
+            <div class="card">
+                <header class="card-header">
+                    <p class="card-header-title">{{blog.title}}</p> 
+                    
+                </header>
+                <div class="card-content">
+                    <div class="content">
+                        {{blog.content}}
+                        <br>
+                        <br>
+                        Posted by {{blog.user.name}} on <time>{{blog.date}}</time>
+                    </div>
+                </div>
+            </div>
+            <br/>
+            {{/each}}
+        </section>
+    </div>
+
+</div>
